Add FPIrVisitor class to compile functions
The FPIrVisitor class was added to facilitate the compilation of functions in the GenericX compiler. This class extends the IrElementVisitorVoid class, enabling the recursive visiting of IR tree nodes. During the traversal, function details are extracted and logged as warnings, helping in debugging and development. Moreover, the logger and the logging tag can be set upon instantiation, providing flexibility in logging settings.